Private set color without control side-effects.
Fixes bug where a black color could not have its hue and saturation set. While logically it makes no sense to do so, and we're not saving that data, we do want the placebo crosshair movement to keep the UI responsive. And so got rid of redundancy in color setting. There's now a private color setter that makes the UI consistent in terms of color, but does not force the graphical selectors into their place.